Реализация Entity

+++ DeliveryApp.Core/Domain/Model/SharedKernel/Location.cs
@@ -6,15 +6,15 @@ namespace DeliveryApp.Core.Domain.Model.SharedKernel
 {
     public class Location : ValueObject
     {
-        public const byte Min = 1;
-        public const byte Max = 10;
+        public const int Min = 1;
+        public const int Max = 10;
 
         [ExcludeFromCodeCoverage]
         private Location()
         {
         }
 
-        private Location(byte x, byte y)
+        private Location(int x, int y)
             : this()
         {
             X = x;
@@ -24,12 +24,12 @@ namespace DeliveryApp.Core.Domain.Model.SharedKernel
         /// <summary>
         /// Координата X
         /// </summary>
-        public byte X { get; }
+        public int X { get; }
 
         /// <summary>
         /// Координата Y
         /// </summary>
-        public byte Y { get; }
+        public int Y { get; }
 
         /// <summary>
         /// Создание объекта Location
@@ -39,7 +39,7 @@ namespace DeliveryApp.Core.Domain.Model.SharedKernel
         /// <returns>
         /// Объект Location
         /// </returns>
-        public static Result<Location, Error> Create(byte x, byte y)
+        public static Result<Location, Error> Create(int x, int y)
         {
             if (x is < Min or > Max)
             {
@@ -63,16 +63,16 @@ namespace DeliveryApp.Core.Domain.Model.SharedKernel
         public static Result<Location, Error> CreateRandom()
         {
             var random = new Random();
-            var x = (byte)random.Next(Min, Max + 1);
-            var y = (byte)random.Next(Min, Max + 1);
+            var x = random.Next(Min, Max + 1);
+            var y = random.Next(Min, Max + 1);
             return new Location(x, y);
         }
 
-        public byte DistanceTo(Location location)
+        public int DistanceTo(Location location)
         {
-            var xDistance = (byte)Math.Abs(location.X - X);
-            var yDistance = (byte)Math.Abs(location.Y - Y);
-            return (byte)(xDistance + yDistance);
+            var xDistance = Math.Abs(location.X - X);
+            var yDistance = Math.Abs(location.Y - Y);
+            return (xDistance + yDistance);
         }
 
         /// <summary>
